+++ shark/src/main/java/com/zx/shark/model/UserBook.java
@@ -1,27 +1,28 @@
 package com.zx.shark.model;
 
 public class UserBook {
-    private int user_id;
-    private int book_id;
+    private Long user_id;
+    private Long book_id;
 
-    public UserBook(int user_id, int book_id) {
+
+    public UserBook(Long user_id, Long book_id) {
         this.user_id = user_id;
         this.book_id = book_id;
     }
 
-    public int getUser_id() {
+    public Long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(Long user_id) {
         this.user_id = user_id;
     }
 
-    public int getBook_id() {
+    public Long getBook_id() {
         return book_id;
     }
 
-    public void setBook_id(int book_id) {
+    public void setBook_id(Long book_id) {
         this.book_id = book_id;
     }
 }
