+++ shark/src/main/java/com/zx/shark/service/BookService.java
@@ -6,8 +6,8 @@ import java.util.List;
 public interface BookService {
     List<Book> findAllBook();
     void removeBook(List<Integer> ids);
-    void borrowBook(int book_id);
-    void cancelBorrow(int book_id);
+    void borrowBook(Long book_id);
+    void cancelBorrow(Long book_id);
     List<Book> selectNeedBook(Book book);
     int countBooks();
